pyproject.toml
[project]
name = "simple-fl"
version = "0.1.0"
description = "A Simple and Professional Federated Learning Framework"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Simple-FL Contributors"}
]
keywords = ["federated-learning", "machine-learning", "deep-learning", "pytorch"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies
dependencies = [
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "numpy>=1.19.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=0.24.0",
    "h5py>=3.1.0",
    "PyYAML>=5.4.0",
    "tqdm>=4.62.0",
    "jsonargparse>=4.0.0",
]

# Optional dependencies (can be installed with: pip install simple-fl[tracking,dev,all])
[project.optional-dependencies]
tracking = [
    "wandb>=0.12.0",
]
visualization = [
    "matplotlib>=3.4.0",
]
notebooks = [
    "jupyter>=1.0.0",
]
database = [
    "pymysql>=1.0.0",
]
dev = [
    "pytest>=6.2.0",
    "black>=21.0",
    "flake8>=3.9.0",
]
all = [
    "wandb>=0.12.0",
    "matplotlib>=3.4.0",
    "jupyter>=1.0.0",
    "pymysql>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/tao-shen/simple-FL"
Repository = "https://github.com/tao-shen/simple-FL"
Issues = "https://github.com/tao-shen/simple-FL/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# Tool configurations
[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

